Add files via upload
This image was generated from using the data of one company (it was chosen from a list of companies where the employees number had a standard deviation of 300 or less to avoid errors or voids in the data). We constructed a dataset with their data regarding the close value of the stock market and the number of employees and the return space (Rt). This is what our dataset looks like for the company with the highest standard deviation in number of employees in our data.